refactor: extract PortfolioTablePrimitives from PortfolioUnifiedTable
Extract CompactInput, MetricValue, WarningMarker, and shared formatting helpers from PortfolioUnifiedTable.tsx into PortfolioTablePrimitives.tsx. Pure refactor — zero behavior change. Shared components will be reused by MobilePortfolioCard in subsequent tickets. Refs: AAV-1115, AAV-1116
